move time_count to utilities file
